HID: steam: add led device and extra key for Steam Button
The wireless Steam Controller has an LED and its brightness can be set,
so add the led class/device.
The user with want to add a special handler for the Steam Button to
work as an exit when another process has a raw connection to the
controller. Therefore add an extra BTN code for it when hidraw is
open.